# hub.read speed up (>6X)