--- a/hub/core/sample.py
+++ b/hub/core/sample.py
@@ -1,5 +1,5 @@
 # type: ignore
-from hub.core.compression import compress_array
+from hub.core.compression import compress_array, verify_compressed_file
 from hub.util.exceptions import CorruptedSampleError
 import numpy as np
 import pathlib
@@ -45,7 +45,7 @@ class Sample:
             self._array = None
             self._read_meta()
             if verify:
-                self._verify()
+                verify_compressed_file(self.path, self.compression)
 
         if array is not None:
             self.path = None
@@ -132,67 +132,6 @@ class Sample:
         self.shape, self._typestr = Image._conv_type_shape(img)
         self.dtype = np.dtype(self._typestr).name
         self.compression = img.format.lower()
-
-    def _verify(self):
-        try:
-            if self.compression == "png":
-                self._verify_png()
-            elif self.compression == "jpeg":
-                self._verify_jpeg()
-            else:
-                self._fast_decompress()
-        except Exception as e:
-            raise CorruptedSampleError(self.compression)
-
-    def _verify_png(self):
-        img = Image.open(self.path)
-        img.verify()
-        img.close()
-
-    def _verify_jpeg(self):
-        # See: https://dev.exiv2.org/projects/exiv2/wiki/The_Metadata_in_JPEG_files#2-The-metadata-structure-in-JPEG
-        with open(self.path, "r+b") as f:
-            mm = mmap.mmap(f.fileno(), 0)
-            soi = f.read(2)
-            # Start of Image
-            assert soi == b"\xff\xd8"
-
-            # Look for Baseline DCT marker
-            sof_idx = mm.find(b"\xff\xc0", 2)
-            if sof_idx == -1:
-                # Look for Progressive DCT marker
-                sof_idx = mm.find(b"\xff\xc2", 2)
-                if sof_idx == -1:
-                    raise Exception()  # Caught by verify
-            f.seek(sof_idx + 2)
-            length = int.from_bytes(f.read(2), "big")
-            f.seek(length - 2, 1)
-            definition_start = f.read(2)
-            assert definition_start in [
-                b"\xff\xc4",
-                b"\xff\xdb",
-                b"\xff\xdd",
-            ]  # DHT, DQT, DRI
-
-            # TODO this check is too slow
-            assert mm.find(b"\xff\xd9") != -1  # End of Image
-
-    def _fast_decompress(self):
-        img = Image.open(self.path)
-        img.load()
-        if img.mode == 1:
-            args = ("L",)
-        else:
-            args = (img.mode,)
-        enc = Image._getencoder(img.mode, "raw", args)
-        enc.setimage(img.im)
-        bufsize = max(65536, img.size[0] * 4)
-        while True:
-            l, s, d = enc.encode(bufsize)
-            if s:
-                break
-        if s < 0:
-            raise Exception()  # caught by _verify()
 
     def __str__(self):
         if self.is_lazy: