… hub/core/tests/test_compression.py
… hub/integrations/pytorch/pytorch_old.py
… hub/integrations/tensorflow.py
… hub/integrations/tests/test_pytorch.py
… hub/integrations/tests/test_tensorflow.py
… hub/tests/dummy_data/compressed_images/corrupt_jpeg.jpeg
… hub/tests/dummy_data/compressed_images/corrupt_png.png
… hub/tests/path_fixtures.py
… hub/util/casting.py
… hub/util/exceptions.py
… hub/util/tests/test_read.py