## 🚀 🚀 Pull Request

### Checklist:

- [x]  [My code follows the style guidelines of this project](https://www.notion.so/activeloop/Engineering-Guidelines-d6e502306d0e4133a8ca507516d1baab) and the [Contributing document](https://github.com/activeloopai/Hub/blob/release/2.0/CONTRIBUTING.md)
- [x]  I have commented my code, particularly in hard-to-understand areas
- [x]  I have kept the `coverage-rate` up
- [x]  I have performed a self-review of my own code and resolved any problems
- [x]  I have checked to ensure there aren't any other open [Pull Requests](https://github.com/activeloopai/Hub/pulls) for the same change
- [x]  I have described and made corresponding changes to the relevant documentation
- [x]  New and existing unit tests pass locally with my changes


### Changes

- Don't decompress the whole sample to get compression format duh!
- Accessing shape and dtype wont require decompression either

Test script:

https://github.com/activeloopai/Hub/blob/main/hub/tests/dummy_data/compressed_images/cat.jpeg

Test image:

https://github.com/activeloopai/Hub/blob/main/hub/tests/dummy_data/compressed_images/cat.jpeg

On `main`:

```sh
PS C:\Users\fariz\Desktop> python test.py
1309.9956512451172 MBs uploaded in 74.37326073646545 seconds. Speed = 17.613798807167562 MB/s)
```

On `fr_read_speedup`

```sh
PS C:\Users\fariz\Desktop> python test.py
1309.9956512451172 MBs uploaded in 12.311607837677002 seconds. Speed = 106.4032958584142 MB/s)
```